+++ genvalidity/src/Data/GenValidity.hs
@@ -180,6 +180,9 @@ instance (GenUnchecked a, GenUnchecked b) => GenUnchecked (a, b) where
             a <- resize r genUnchecked
             b <- resize s genUnchecked
             return (a, b)
+    shrinkUnchecked (a, b) = ((,) <$> shrinkUnchecked a <*> shrinkUnchecked b)
+      ++ [ (a', b) | a' <- shrinkUnchecked a ]
+      ++ [ (a, b') | b' <- shrinkUnchecked b ]
 
 instance (GenValid a, GenValid b) => GenValid (a, b) where
     genValid =
@@ -188,6 +191,9 @@ instance (GenValid a, GenValid b) => GenValid (a, b) where
             a <- resize r genValid
             b <- resize s genValid
             return (a, b)
+    shrinkValid (a, b) = ((,) <$> shrinkValid a <*> shrinkValid b)
+      ++ [ (a', b) | a' <- shrinkValid a ]
+      ++ [ (a, b') | b' <- shrinkValid b ]
 
 instance (GenInvalid a, GenInvalid b) => GenInvalid (a, b) where
     genInvalid =
@@ -204,9 +210,13 @@ instance (GenInvalid a, GenInvalid b) => GenInvalid (a, b) where
 
 instance (GenUnchecked a, GenUnchecked b) => GenUnchecked (Either a b) where
     genUnchecked = oneof [Left <$> genUnchecked, Right <$> genUnchecked]
+    shrinkUnchecked (Left a) = Left <$> shrinkUnchecked a
+    shrinkUnchecked (Right b) = Right <$> shrinkUnchecked b
 
 instance (GenValid a, GenValid b) => GenValid (Either a b) where
     genValid = oneof [Left <$> genValid, Right <$> genValid]
+    shrinkValid (Left a) = Left <$> shrinkValid a
+    shrinkValid (Right b) = Right <$> shrinkValid b
 
 -- | This instance ensures that the generated tupse contains at least one invalid element. The other element is unchecked.
 instance (GenInvalid a, GenInvalid b) => GenInvalid (Either a b) where
@@ -221,6 +231,10 @@ instance (GenUnchecked a, GenUnchecked b, GenUnchecked c) =>
             b <- resize s genUnchecked
             c <- resize t genUnchecked
             return (a, b, c)
+    shrinkUnchecked (a, b, c) =
+        [ (a', b', c')
+        | (a', (b', c')) <- shrinkUnchecked (a, (b, c))
+        ]
 
 instance (GenValid a, GenValid b, GenValid c) => GenValid (a, b, c) where
     genValid =
@@ -230,6 +244,10 @@ instance (GenValid a, GenValid b, GenValid c) => GenValid (a, b, c) where
             b <- resize s genValid
             c <- resize t genValid
             return (a, b, c)
+    shrinkValid (a, b, c) =
+        [ (a', b', c')
+        | (a', (b', c')) <- shrinkValid (a, (b, c))
+        ]
 
 -- | This instance ensures that the generated triple contains at least one invalid element. The other two are unchecked.
 instance (GenInvalid a, GenInvalid b, GenInvalid c) =>
@@ -262,6 +280,10 @@ instance (GenUnchecked a, GenUnchecked b, GenUnchecked c, GenUnchecked d) =>
             c <- resize t genUnchecked
             d <- resize u genUnchecked
             return (a, b, c, d)
+    shrinkUnchecked (a, b, c, d) =
+        [ (a', b', c', d')
+        | (a', (b', (c', d'))) <- shrinkUnchecked (a, (b, (c, d)))
+        ]
 
 instance (GenValid a, GenValid b, GenValid c, GenValid d) =>
          GenValid (a, b, c, d) where
@@ -273,6 +295,10 @@ instance (GenValid a, GenValid b, GenValid c, GenValid d) =>
             c <- resize t genValid
             d <- resize u genValid
             return (a, b, c, d)
+    shrinkValid (a, b, c, d) =
+        [ (a', b', c', d')
+        | (a', (b', (c', d'))) <- shrinkValid (a, (b, (c, d)))
+        ]
 
 -- | This instance ensures that the generated triple contains at least one invalid element. The other two are unchecked.
 instance (GenInvalid a, GenInvalid b, GenInvalid c, GenInvalid d) =>
@@ -314,6 +340,10 @@ instance (GenUnchecked a, GenUnchecked b, GenUnchecked c, GenUnchecked d, GenUnc
             d <- resize u genUnchecked
             e <- resize v genUnchecked
             return (a, b, c, d, e)
+    shrinkUnchecked (a, b, c, d, e) =
+        [ (a', b', c', d', e')
+        | (a', (b', (c', (d', e')))) <- shrinkUnchecked (a, (b, (c, (d, e))))
+        ]
 
 instance (GenValid a, GenValid b, GenValid c, GenValid d, GenValid e) =>
          GenValid (a, b, c, d, e) where
@@ -326,6 +356,10 @@ instance (GenValid a, GenValid b, GenValid c, GenValid d, GenValid e) =>
             d <- resize u genValid
             e <- resize v genValid
             return (a, b, c, d, e)
+    shrinkValid (a, b, c, d, e) =
+        [ (a', b', c', d', e')
+        | (a', (b', (c', (d', e')))) <- shrinkValid (a, (b, (c, (d, e))))
+        ]
 
 -- | This instance ensures that the generated triple contains at least one invalid element. The other two are unchecked.
 instance (GenInvalid a, GenInvalid b, GenInvalid c, GenInvalid d, GenInvalid e) =>
@@ -368,9 +402,14 @@ instance (GenInvalid a, GenInvalid b, GenInvalid c, GenInvalid d, GenInvalid e) 
 
 instance GenUnchecked a => GenUnchecked (Maybe a) where
     genUnchecked = oneof [pure Nothing, Just <$> genUnchecked]
+    shrinkUnchecked Nothing = []
+    shrinkUnchecked (Just a) = Nothing : (Just <$> shrinkUnchecked a)
+
 
 instance GenValid a => GenValid (Maybe a) where
     genValid = oneof [pure Nothing, Just <$> genValid]
+    shrinkValid Nothing = []
+    shrinkValid (Just a) = Nothing : (Just <$> shrinkValid a)
 
 instance GenInvalid a => GenInvalid (Maybe a) where
     genInvalid = Just <$> genInvalid
@@ -386,6 +425,7 @@ instance GenUnchecked a => GenUnchecked (NonEmpty a) where
 
 instance GenValid a => GenValid (NonEmpty a) where
     genValid = (:|) <$> genValid <*> genValid
+    shrinkValid (v :| vs) = [ e :| es | (e, es) <- shrinkValid (v, vs)]
 
 instance GenInvalid a => GenInvalid (NonEmpty a) where
     genInvalid = sized $ \n -> do
@@ -400,6 +440,7 @@ instance GenInvalid a => GenInvalid (NonEmpty a) where
 -- them.
 instance GenValid a => GenValid [a] where
     genValid = genListOf genValid
+    shrinkValid = shrinkList shrinkValid
 
 -- | This instance ensures that the generated list contains at least one element
 -- that satisfies 'isInvalid'. The rest is unchecked.